// src/modules/orders/interface/order.interface.ts
import { Document } from 'mongoose';
export interface IOrder extends Document {
  readonly _id: string;
  readonly id: number;
  readonly parent_id: number;
  readonly status: string;
  readonly currency: string;
  readonly version: string;
  readonly prices_include_tax: boolean;
  readonly date_created: [];
  readonly date_modified: [];
  readonly discount_total: string;
  readonly discount_tax: string;
  readonly shipping_total: string;
  readonly shipping_tax: string;
  readonly cart_tax: string;
  readonly total: string;
  readonly customer_id: number;
  readonly order_key: string;
  readonly billing: [];
  readonly shipping: [];
  readonly payment_method: string;
  readonly payment_method_title: string;
  readonly transaction_id: string;
  readonly customer_ip_address: string;
  readonly customer_user_agent: string;
  readonly created_via: string;
  readonly customer_note: string;
  readonly date_completed: [];
  readonly date_paid: string;
  readonly cart_hash: string;
  readonly order_stock_reduced: string;
  readonly download_permissions_granted: string;
  readonly new_order_email_sent: string;
  readonly recorded_sales: string;
  readonly recorded_coupon_usage_counts: string;
  readonly number: string;
  readonly meta_data: [];
  readonly line_items: [];
  readonly tax_lines: [];
  readonly shipping_lines: [];
  readonly fee_lines: [];
  readonly coupon_lines: [];
  readonly order_notes: [];
  readonly easy_post_data: [];
  readonly subscription_orders: [];
  readonly refunds: [];
  readonly payment_url: string;
  readonly is_editable: boolean;
  readonly needs_payment: boolean;
  readonly needs_processing: boolean;
  readonly date_created_gmt: string;
  readonly date_modified_gmt: string;
  readonly date_completed_gmt: [];
  readonly date_paid_gmt: string;
  readonly currency_symbol: string;
  readonly _links: [];
}